+++ ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: elehtora <elehtora@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/08 13:09:05 by elehtora          #+#    #+#             */
-/*   Updated: 2022/02/20 16:02:10 by elehtora         ###   ########.fr       */
+/*   Updated: 2022/02/20 17:48:27 by elehtora         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ char	*ft_strtrim(char const *s)
 		return (NULL);
 	start = get_start(s);
 	tail = get_tail(start);
-	trimmed = ft_strnew(tail - start + 1);
+	trimmed = ft_strnew((size_t)(tail - start + 1));
 	if (trimmed == NULL)
 		return (NULL);
 	result_ptr = trimmed;
